+++ src/crunch.c
@@ -35,7 +35,7 @@ along with LUX.  If not, see <http://www.gnu.org/licenses/>.
 /* commented out because already defined in .h file - LS 30apr97 */
 Byte bits[8]={1,2,4,8,16,32,64,128};
 Int	crunch_bits;
-float	crunch_bpp;
+double	crunch_bpp;
 Int	crunch_slice = 5;
 static Int	crunch_run_flag=0;
