+++ src/decomp.c
@@ -56,13 +56,13 @@ k++; qd += (nd + 1); }				/*end of outer loop */
 return 1;			/* added LS 5feb97 */
 }
 /*--------------------------------------------------------------------------*/
-Int f_decomp(float *x, Int n, Int nd)
+Int f_decomp(double *x, Int n, Int nd)
 /*translated from fortran anadecomp.for which may be easier to follow since
 	it uses subscripts rather than the pointers used here */
 /*no pivoting in this version !, so diagonals must be != 0 */
 {
-  register	float	sum1, sum2, *p1, *p2, *p3, *p4;
-  float		*qd, * q2, *q1, div;
+  register	double	sum1, sum2, *p1, *p2, *p3, *p4;
+  double		*qd, * q2, *q1, div;
   Int		nq, mq, lq, k;
 
   p1 = x;
@@ -90,10 +90,10 @@ Int f_decomp(float *x, Int n, Int nd)
 return 1;			/* added LS 5feb97 */
 }
 /*--------------------------------------------------------------------------*/
-Int f_solve(float *a, float *b, Int n, Int nd)
+Int f_solve(double *a, double *b, Int n, Int nd)
 {
-  register	float	sum, *p1, *p2;
-  float		*qd, *q1;
+  register	double	sum, *p1, *p2;
+  double		*qd, *q1;
   Int		nq, mq, k;
 
 /*printf("in f_solve\n");*/
